--- a/test/benchmark.js
+++ b/test/benchmark.js
@@ -422,6 +422,11 @@ describe("benchmark", () => {
       // Now, normalize file lists before comparing.
       yarnFiles = yarnFiles.sort();
 
+      // This diff dependency is expected to **stay** in place because we test
+      // forcing versions to prevent flattening.
+      const NESTED_DIFF = "functions/base/node_modules/diff/";
+
+
       const NPM_NORMS = {
         // Duplicated in yarn.
         // eslint-disable-next-line max-len
@@ -430,14 +435,15 @@ describe("benchmark", () => {
         "functions/base/node_modules/cookie/": "node_modules/express/node_modules/cookie/",
         "functions/base/node_modules/send/node_modules/ms/":
           "node_modules/debug/node_modules/ms/",
-        // Hoist everything to root (which is what yarn should do).
+        // Hoist everything to root (which is what yarn should do), except for
+        // `/diff/` which we've engineered to stay in place...
         "functions/base/node_modules/": "node_modules/",
         "lib/camel/node_modules/": "node_modules/"
       };
       npmFiles = npmFiles
         .map((dep) => {
           for (const norm of Object.keys(NPM_NORMS)) {
-            if (dep.startsWith(norm)) {
+            if (dep.startsWith(norm) && !dep.startsWith(NESTED_DIFF)) {
               return NPM_NORMS[norm] === null ? null : dep.replace(norm, NPM_NORMS[norm]);
             }
           }