Attempts to address issue #53 

In order to fix this issue the following changes were made:

- reverse the order of globs so that dependency includes come after service/function-level includes
- adds a test case for `broad negative include with include re-adding in`
- provides the `depInclude` array to `filterFiles` and match depIncludes after matching all service/function includes/excludes

Some stats on the artifacts:
```
// before changes
$ du -sh ./test/packages/monorepo/yarn/.serverless/*
860K    ./test/packages/monorepo/yarn/.serverless/another.zip
696K    ./test/packages/monorepo/yarn/.serverless/base.zip
// after changes
$ du -sh ./test/packages/monorepo/yarn/.serverless/*
860K    ./test/packages/monorepo/yarn/.serverless/another.zip
916K    ./test/packages/monorepo/yarn/.serverless/base.zip
// diff package compressed (the dependency that was missing from the artifact created on master)
$ du -sh ./diff.zip
204K    ./diff.zip
```

# Benchmarks

# Before Changes


## Benchmark: System Information
* os:   `darwin 18.5.0 x64`
* node: `v10.15.0`
* yarn: `1.16.0`
* npm:  `6.4.1`

## Benchmark: Timed Packages
| Scenario     | Mode | Type     |  Time |      vs Base |
| :----------- | :--- | :------- | ----: | -----------: |
| simple       | yarn | jetpack  |  1802 | **-66.08 %** |
| simple       | yarn | baseline |  5312 |              |
| simple       | npm  | jetpack  |  2005 | **-63.47 %** |
| simple       | npm  | baseline |  5488 |              |
| complex      | yarn | jetpack  |  3670 | **-83.22 %** |
| complex      | yarn | baseline | 21875 |              |
| complex      | npm  | jetpack  |  3542 | **-84.23 %** |
| complex      | npm  | baseline | 22467 |              |
| individually | yarn | jetpack  |  2116 | **-72.32 %** |
| individually | yarn | baseline |  7645 |              |
| individually | npm  | jetpack  |  2010 | **-72.93 %** |
| individually | npm  | baseline |  7424 |              |
| huge         | yarn | jetpack  |  2867 | **-85.12 %** |
| huge         | yarn | baseline | 19262 |              |
| huge         | npm  | jetpack  |  3008 | **-85.12 %** |
| huge         | npm  | baseline | 20219 |              |

## Benchmark: Other Packages
| Scenario | Mode | Type     | Time |    vs Base |
| :------- | :--- | :------- | ---: | ---------: |
| monorepo | yarn | jetpack  | 1803 |            |
| monorepo | npm  | jetpack  | 2366 |            |
| webpack  | yarn | jetpack  | 1452 | **3.27 %** |
| webpack  | yarn | baseline | 1406 |            |
| webpack  | npm  | jetpack  | 1605 | **9.86 %** |
| webpack  | npm  | baseline | 1461 |            |

# After Changes

## Benchmark: System Information
* os:   `darwin 18.5.0 x64`
* node: `v10.15.0`
* yarn: `1.16.0`
* npm:  `6.4.1`

## Benchmark: Timed Packages
| Scenario     | Mode | Type     |  Time |      vs Base |
| :----------- | :--- | :------- | ----: | -----------: |
| simple       | yarn | jetpack  |  2042 | **-60.70 %** |
| simple       | yarn | baseline |  5196 |              |
| simple       | npm  | jetpack  |  1806 | **-66.98 %** |
| simple       | npm  | baseline |  5469 |              |
| complex      | yarn | jetpack  |  3623 | **-82.90 %** |
| complex      | yarn | baseline | 21192 |              |
| complex      | npm  | jetpack  |  3627 | **-84.51 %** |
| complex      | npm  | baseline | 23415 |              |
| individually | yarn | jetpack  |  2343 | **-69.33 %** |
| individually | yarn | baseline |  7640 |              |
| individually | npm  | jetpack  |  2094 | **-73.82 %** |
| individually | npm  | baseline |  7997 |              |
| huge         | yarn | jetpack  |  2155 | **-87.85 %** |
| huge         | yarn | baseline | 17732 |              |
| huge         | npm  | jetpack  |  2081 | **-88.76 %** |
| huge         | npm  | baseline | 18514 |              |

## Benchmark: Other Packages
| Scenario | Mode | Type     | Time |      vs Base |
| :------- | :--- | :------- | ---: | -----------: |
| monorepo | yarn | jetpack  | 2543 |              |
| monorepo | npm  | jetpack  | 1946 |              |
| webpack  | yarn | jetpack  | 2000 |  **16.21 %** |
| webpack  | yarn | baseline | 1721 |              |
| webpack  | npm  | jetpack  | 1593 | **-11.99 %** |
| webpack  | npm  | baseline | 1810 |              |


there does not seem to be a statistically significant difference in performance after the changes