@pdeona 2019-07-21
updated, now passing all tests and including function dependencies correctly