@ryan-roemer 2019-07-21
Wow! What a lot of great work!

One thing we need to keep in mind is the arguably top goal of this project -- being a drop-in replacement for `serverless` packaging, which means abiding by the globby + nanomatch logic (or an equivalent) to produce identical packages outside of bespoke functionality like monorepo support (using `roots` or `base`, etc.)

This branch appears to fail `yarn benchmark:test` for 3 scenarios. One is `monorepo  has same npm and yarn package contents` and can be ignored as the fixture changes are causing that from my upstream branch and my upstream branch fails too -- something we'll fix at the end. The other two failures are `benchmark > baseline vs jetpack > individually/yarn` and `benchmark > baseline vs jetpack > individually/npm` and I believe those are legitimate failures that indicate this branch now deviates in globbing behavior from default `serverless` packaging. Could you take a look there and see if that is the case, and if so, if we can get this plugin back in line?

I've been giving a little thought to the overall problem of "do dep includes before or after normal includes in `globby()`" and I'm not sure if the solution is addressable straight up while maintaining compatibility with `severless`. It may be that we just add something like an "extra" `include`/`exclude` (named something different) that then comes _after_ service + function include/exclude like normal for a bit more fine-grained control... Or some other "out of the box solution".

... at any rate, just some random late night weekend thoughts. I'm planning on digging in more Monday afternoon when I get some time and maybe we can touch base then if there's not some easy solution to the above that becomes apparent...