@pdeona 2019-07-21
i probably should have run these tests at some point! oops :P will see if i can get these `individually` cases passing tomorrow morning. thanks for the idea as well, will probably need to rethink my approach if this has broken existing functionality.