@pdeona 2019-07-21
@ryan-roemer working from your comment about globby i was able to come up with a much smaller overall changeset that still solves the monorepo dependency issue, and it looks like all the integration suites are passing, with the exception of the monorepo one you mentioned. let me know if you see any other issues here, otherwise i'll see if i can get that monorepo case passing later.